PEEK-692 fix alembic revision autogenerate
"alembic revision --autogenerate" raises
sqlalchemy.exc.NoSuchTableError.

The cause is that when alembic runs the mirgration, it fails to
filter other schmeas even if there is a includeFilter specified.

The fix is to configure alembic to compare the changes with each
schema and make transcations for each schema instead of doing a
transcation for all changes.